Refactor: 패키지 구조가 1개로 변경되어 shared패키지 제거
