+++ gathering-service/src/main/java/com/sparta/moim/gathering/gathering/application/service/GatheringValidationService.java
@@ -1,8 +1,8 @@
 package com.sparta.moim.gathering.gathering.application.service;
 
 import com.sparta.moim.gathering.gathering.domain.repository.GatheringValidationRepository;
-import com.sparta.moim.gathering.shared.error.code.GatheringCode;
-import com.sparta.moim.gathering.shared.error.exception.GatheringException;
+import com.sparta.moim.gathering.gathering.application.code.GatheringCode;
+import com.sparta.moim.gathering.gathering.application.exception.GatheringException;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
